Initial GUI
Home controller passes a MonthyExpenseModel to the Index view for displaying the current month and the month's expenses
Home controller serves the page for adding expenses manually
Home controller processes newly added expenses
Removed privacy pages since we are not publishing the web application
Added test cases for adding expenses to this month and removing test expenses
Added month enum to ExpenseModel for representing the month with the name of the month (January, February, etc)
Created MonhtlyExpenseModel to contain the currently selected date and a list of expenses. This allows selecting months without expenses for the main view
Em ignored the Forms application inserted into the project so the project will compile
Changed expense repo to sort expenses by date. They are not sorted in the SQL database and should be sorted before being presented to the user
Created AddExpense form for adding expenses manually
Created PartialView ExpensesByDate for ajax loading of user expenses
Modified Index view to show the ExpensesByDate partial view
Added javascript function to Index view for selecting the month to load via ajax
Modified the layout to be more minimalistic
Modified test results to show passes in green and failures in red
Created plus icon for adding expenses